Only show Okular actions in the menu when Okular is installed on linux
