+++ src/nl/hannahsten/texifyidea/action/group/OkularActionGroup.kt
@@ -2,11 +2,11 @@ package nl.hannahsten.texifyidea.action.group
 
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.actionSystem.DefaultActionGroup
+import nl.hannahsten.texifyidea.run.okular.isOkularAvailable
 
 class OkularActionGroup : DefaultActionGroup() {
 
-    // TODO should be isOkularAvailable()
-    override fun canBePerformed(context: DataContext) = true
+    override fun canBePerformed(context: DataContext) = isOkularAvailable()
 
     override fun hideIfNoVisibleChildren(): Boolean = true
 }
