+++ src/nl/hannahsten/texifyidea/run/evince/EvinceConversation.kt
@@ -2,6 +2,7 @@ package nl.hannahsten.texifyidea.run.evince
 
 import com.intellij.openapi.util.SystemInfo
 import nl.hannahsten.texifyidea.TeXception
+import nl.hannahsten.texifyidea.run.runCommand
 import org.freedesktop.dbus.connections.impl.DBusConnection
 import org.freedesktop.dbus.errors.NoReply
 import org.gnome.evince.Daemon
@@ -21,29 +22,7 @@ fun isEvinceAvailable(): Boolean {
 
     // Find out whether Evince is installed and in PATH, otherwise we can't use it
     val output = "which evince".runCommand()
-    return output?.contains("evince") ?: false
-}
-
-/**
- * Run a command in the terminal.
- *
- * @return The output of the command or null if an exception was thrown.
- */
-fun String.runCommand(): String? {
-    return try {
-        val parts = this.split("\\s".toRegex())
-        val proc = ProcessBuilder(*parts.toTypedArray())
-                .redirectOutput(ProcessBuilder.Redirect.PIPE)
-                .redirectError(ProcessBuilder.Redirect.PIPE)
-                .start()
-
-        // Timeout value
-        proc.waitFor(10, TimeUnit.SECONDS)
-        proc.inputStream.bufferedReader().readText()
-    } catch (e: IOException) {
-        e.printStackTrace()
-        null
-    }
+    return output?.contains("/evince") ?: false
 }
 
 /**
